the info package has been renamed to locator;
