+++ app/src/main/java/com/team/mamba/atlascalendar/userInterface/dashBoard/announcements/AnnouncementsFragment.java
@@ -24,7 +24,7 @@ import com.team.mamba.atlascalendar.userInterface.dashBoard._container_activity.
 import com.team.mamba.atlascalendar.userInterface.dashBoard.announcements.create_announcement.CreateAnnouncementFragment;
 import com.team.mamba.atlascalendar.userInterface.dashBoard.contacts.ContactsFragment;
 import com.team.mamba.atlascalendar.userInterface.dashBoard.crm.main.CrmFragment;
-import com.team.mamba.atlascalendar.userInterface.dashBoard.info.InfoFragment;
+import com.team.mamba.atlascalendar.userInterface.dashBoard.locator.LocatorFragment;
 import com.team.mamba.atlascalendar.utils.AppConstants;
 import com.team.mamba.atlascalendar.utils.ChangeFragments;
 
@@ -160,7 +160,7 @@ public class AnnouncementsFragment extends BaseFragment<AnnouncementsLayoutBindi
     public void onInfoClicked() {
 
         FragmentManager manager = getBaseActivity().getSupportFragmentManager();
-        ChangeFragments.replaceFromBackStack(InfoFragment.newInstance(), manager, "Info", null);
+        ChangeFragments.replaceFromBackStack(LocatorFragment.newInstance(), manager, "Info", null);
     }
 
     @Override
