+++ app/src/main/java/com/team/mamba/atlascalendar/userInterface/dashBoard/crm/main/CrmFragment.java
@@ -46,7 +46,7 @@ import com.team.mamba.atlascalendar.userInterface.dashBoard.crm.filter_list.CrmF
 import com.team.mamba.atlascalendar.userInterface.dashBoard.crm.selected_crm.SelectedCrmFragment;
 import com.team.mamba.atlascalendar.userInterface.dashBoard._container_activity.DashBoardActivity;
 import com.team.mamba.atlascalendar.userInterface.dashBoard._container_activity.DashBoardActivityNavigator;
-import com.team.mamba.atlascalendar.userInterface.dashBoard.info.InfoFragment;
+import com.team.mamba.atlascalendar.userInterface.dashBoard.locator.LocatorFragment;
 import com.team.mamba.atlascalendar.utils.AppConstants;
 import com.team.mamba.atlascalendar.utils.ChangeFragments;
 
@@ -241,7 +241,7 @@ public class CrmFragment extends BaseFragment<CrmLayoutBinding, CrmViewModel>
     public void onInfoClicked() {
 
         FragmentManager manager = getBaseActivity().getSupportFragmentManager();
-        ChangeFragments.replaceFromBackStack(InfoFragment.newInstance(), manager, "Info", null);
+        ChangeFragments.replaceFromBackStack(LocatorFragment.newInstance(), manager, "Info", null);
     }
 
     @Override
