--- app/src/main/java/com/team/mamba/atlascalendar/userInterface/dashBoard/info/InfoDataModel.java
@@ -1,34 +1,0 @@
-package com.team.mamba.atlascalendar.userInterface.dashBoard.info;
-
-import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.messaging.FirebaseMessaging;
-import com.orhanobut.logger.Logger;
-import com.team.mamba.atlascalendar.data.AppDataManager;
-import com.team.mamba.atlascalendar.data.model.api.fireStore.BusinessProfile;
-import com.team.mamba.atlascalendar.data.model.api.fireStore.CrmNotes;
-import com.team.mamba.atlascalendar.data.model.api.fireStore.UserConnections;
-import com.team.mamba.atlascalendar.data.model.api.fireStore.UserProfile;
-import com.team.mamba.atlascalendar.utils.AppConstants;
-
-import java.text.DecimalFormat;
-import java.util.ArrayList;
-import java.util.Calendar;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-
-import javax.inject.Inject;
-
-public class InfoDataModel {
-
-
-    private AppDataManager dataManager;
-
-    @Inject
-    public InfoDataModel(AppDataManager dataManager) {
-
-        this.dataManager = dataManager;
-    }
-
-
-}
