--- app/src/main/java/com/team/mamba/atlascalendar/userInterface/dashBoard/info/InfoNavigator.java
@@ -1,19 +1,0 @@
-package com.team.mamba.atlascalendar.userInterface.dashBoard.info;
-
-import com.team.mamba.atlascalendar.data.model.api.fireStore.UserConnections;
-
-import java.util.List;
-
-public interface InfoNavigator {
-
-    void onSettingsClicked();
-
-    void handleError(String msg);
-
-    void onCrmClicked();
-
-    void onNotificationsClicked();
-
-    void onContactsClicked();
-
-}
