+++ app/src/main/java/com/team/mamba/atlascalendar/userInterface/dashBoard/locator/LocatorNavigator.java
@@ -1,0 +1,15 @@
+package com.team.mamba.atlascalendar.userInterface.dashBoard.locator;
+
+public interface LocatorNavigator {
+
+    void onSettingsClicked();
+
+    void handleError(String msg);
+
+    void onCrmClicked();
+
+    void onNotificationsClicked();
+
+    void onContactsClicked();
+
+}
