+++ app/src/main/res/layouts/dashBoard/info/layout/info_layout.xml
@@ -14,7 +14,7 @@
 
         <variable
             name = "viewModel"
-            type = "com.team.mamba.atlascalendar.userInterface.dashBoard.info.InfoViewModel"/>
+            type = "com.team.mamba.atlascalendar.userInterface.dashBoard.locator.LocatorViewModel"/>
 
     </data>
