+++ app/src/main/res/layouts/dashBoard/info/layout/recent_activities_list_row.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="viewModel"
-            type="com.team.mamba.atlascalendar.userInterface.dashBoard.info.InfoViewModel" />
+            type="com.team.mamba.atlascalendar.userInterface.dashBoard.locator.LocatorViewModel" />
 
     </data>
