+++ app/src/main/res/layouts/dashBoard/info/layout/user_stats_list_row.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="viewModel"
-            type="com.team.mamba.atlascalendar.userInterface.dashBoard.info.InfoViewModel" />
+            type="com.team.mamba.atlascalendar.userInterface.dashBoard.locator.LocatorViewModel" />
 
     </data>
